🎉  (next-config) NICE-34 prep for multi-site
Few variables to pass from site to config
May continue to fine-tune so stuff like:
 - securityHeaders

And more are separated but this is pretty specific
 at the moment and not something that should be
 published. Morel ike a `next-config` that is localized
 only.

♻️  (next-config) NICE-34 breakout a bit more
