+++ packages/next-config/package.json
@@ -1,0 +1,14 @@
+{
+  "name": "@jeromefitz/next-config",
+  "version": "0.0.0",
+  "private": true,
+  "main": "./next.config.js",
+  "devDependencies": {
+    "@jeromefitz/prettier-config": "1.1.6",
+    "@next/bundle-analyzer": "13.3.1",
+    "@octokit/core": "4.2.0",
+    "@plaiceholder/next": "2.5.0",
+    "fast-json-stable-stringify": "2.1.0",
+    "lodash": "4.17.21"
+  }
+}
