+++ sites/jeromefitzgerald.com/package.json
@@ -40,8 +40,6 @@
     "@jeromefitz/spotify": "3.0.2",
     "@jeromefitz/utils": "2.2.2",
     "@mantine/hooks": "6.0.8",
-    "@octokit/core": "4.2.0",
-    "@plaiceholder/next": "2.5.0",
     "@radix-ui/colors": "0.1.8",
     "@radix-ui/primitive": "1.0.0",
     "@radix-ui/react-accessible-icon": "1.0.2",
@@ -141,9 +139,9 @@
     "zustand": "4.3.7"
   },
   "devDependencies": {
+    "@jeromefitz/next-config": "workspace:*",
     "@jeromefitz/shared": "workspace:*",
     "@jeromefitz/tsconfig": "1.1.3",
-    "@next/bundle-analyzer": "13.3.1",
     "@plaiceholder/tailwindcss": "2.5.0",
     "@tailwindcss/aspect-ratio": "0.4.2",
     "@tailwindcss/typography": "0.5.9",
